Move command to command/mod.rs
Split TrigArgs and friends into another module
